VHE exposure lengths
The following are the latest efforts to create a working modification of
the exposure lengths.  I will be incorporating it into a new python file
that is dedicated to modifying the time bins to mimic those of a VHE
instrument. Unfortunately, I had to refresh my operating system, so I am
also reinstalling python.